:recycle: Extract todo labels widget
